requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
sentence-transformers==2.5.1
torch==2.2.0
numpy==1.26.4
gunicorn==21.2.0
mongoengine==0.24.2
pymongo[srv]==3.12.3
dnspython==2.4.2
certifi==2024.2.2
requests==2.31.0